post schema update relational collection after add or delete document

+++ controllers/song.controller.js
@@ -1,12 +1,16 @@
 const Songs = require("../models/song.model");
+const { uploadFile, deleteFileStream } = require("../config/s3");
 
 module.exports = {
   getAll: async (req, res) => {
     try {
-      const data = await Songs.find({}, "-__v").populate("albums", "name -_id");
+      const data = await Songs.find({}, "-__v")
+        .populate("artist", "name image")
+        .populate("album", "name image");
+
       res.json({
         message: "Success get All Songs",
-        data: data,
+        songs: data,
       });
     } catch (err) {
       console.log(err);
@@ -15,13 +19,13 @@ module.exports = {
   },
   getById: async (req, res) => {
     try {
-      const data = await Songs.findById(req.params.id, "-__v").populate(
-        "albums",
-        "name -_id"
-      );
+      const data = await Songs.findById(req.params.id, "-__v")
+        .populate("album", "name image")
+        .populate("artist", "name image");
+
       res.json({
         message: "Success get All Songs By ID",
-        data: data,
+        songs: data,
       });
     } catch (err) {
       console.log(err);
@@ -31,6 +35,16 @@ module.exports = {
   addSong: async (req, res) => {
     try {
       const data = req.body;
+
+      console.log(req.file);
+
+      // upload file to aws s3
+      if (req.file) {
+        const result = await uploadFile(req.file);
+        console.log(result);
+        data.file = result.key;
+      }
+
       await Songs.create(data);
       res.json({
         message: "Success add Songs",
@@ -43,9 +57,25 @@ module.exports = {
 
   updateSongById: async (req, res) => {
     try {
-      await Songs.updateOne({ _id: req.params.id }, { $set: req.body });
+      const song = (await Songs.findById(req.params.id, "-__v")) || false;
+      const data = req.body;
+
+      if (!song) return res.status(404).json({ messege: "song not found" });
+
+      // check file dan delete file lama
+      if (req.file) {
+        const result = await uploadFile(req.file);
+        console.log(result);
+        data.file = result.key;
+
+        // delete image on aws server
+        await deleteFileStream(song.file);
+      }
+
+      await Songs.updateOne({ _id: req.params.id }, { $set: data });
       res.json({
         message: "Success Update Songs",
+        song: data,
       });
     } catch (err) {
       console.log(err);
@@ -55,7 +85,11 @@ module.exports = {
 
   deletesongById: async (req, res) => {
     try {
+      const song = (await Songs.findById(req.params.id, "-__v")) || false;
+      if (!song) return res.status(404).json({ messege: "song not found" });
+
       await Songs.deleteOne({ _id: req.params.id });
+      await deleteFileStream(song.file);
       res.json({
         message: "Succes delete Songs",
       });
